+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## [1.0.0]
+* Fixed failing validation on password form fields
+* MinLength & MaxLength validation fix
+* Version 1 ready
 ## [0.0.2]
 * Added test for minLength()
 * Updated the readme & info
-## [0.0.1] - TODO: Add release date.
+## [0.0.1]
 * Initial Release
 * TODO (Features):
 * Add Url Validation